start  restructuring toward a single client chapter
