+++ src/lib.rs
@@ -4,7 +4,7 @@ use std::hash::{Hash, Hasher};
 mod c1_state_machine;
 mod c2_blockchain;
 mod c3_consensus;
-mod c4_framework;
+mod c5_client;
 
 // Simple helper to do some hashing.
 fn hash<T: Hash>(t: &T) -> u64 {
